producer/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.11.1
pydantic>=1.8,<1.9
typing_extensions<4.6.0
numpy==1.21.6
thinc==8.0.15
spaCy==3.2.0
